introduced pug for html templating - added cards and nav pages - updated some navbar items
