+++ .jsbeautifyrc
@@ -1,0 +1,13 @@
+ {
+   "allowed_file_extensions": ["htm", "html", "xhtml", "shtml", "xml", "svg", "dust"],
+   "brace_style": "collapse",
+   "end_with_newline": true,
+   "indent_char": " ",
+   "indent_inner_html": false,
+   "indent_scripts": "keep",
+   "indent_size": 2,
+   "max_preserve_newlines": 1,
+   "preserve_newlines": true,
+   "unformatted": ["code", "h1", "h2", "h3", "h4", "h5", "h6"],
+   "wrap_line_length": 0
+ }
