+++ gulpfile.js
@@ -3,8 +3,10 @@ var sass = require('gulp-sass');
 var browserSync = require('browser-sync').create();
 var header = require('gulp-header');
 var cleanCSS = require('gulp-clean-css');
+var pug = require('gulp-pug');
 var rename = require("gulp-rename");
 var uglify = require('gulp-uglify');
+var beautify = require('gulp-html-beautify');
 var pkg = require('./package.json');
 
 // Set the banner content
@@ -46,7 +48,7 @@ gulp.task('minify-css', ['sass'], function() {
 
 // Minify custom JS
 gulp.task('minify-js', function() {
-  return gulp.src('js/sb-admin.js')
+  return gulp.src(['js/**/*.js', '!js/**/*.min.js'])
     .pipe(uglify())
     .pipe(header(banner, {
       pkg: pkg
@@ -55,6 +57,17 @@ gulp.task('minify-js', function() {
       suffix: '.min'
     }))
     .pipe(gulp.dest('js'))
+    .pipe(browserSync.reload({
+      stream: true
+    }))
+});
+
+// Compiles Pug files from /pug into the root folder and beautifies the HTML
+gulp.task('pug', function buildHTML() {
+  return gulp.src('pug/*.pug')
+    .pipe(pug())
+    .pipe(beautify())
+    .pipe(gulp.dest(''))
     .pipe(browserSync.reload({
       stream: true
     }))
@@ -102,7 +115,7 @@ gulp.task('copy', function() {
 })
 
 // Default task
-gulp.task('default', ['sass', 'minify-css', 'minify-js', 'copy']);
+gulp.task('default', ['sass', 'minify-css', 'minify-js', 'pug', 'copy']);
 
 // Configure the browserSync task
 gulp.task('browserSync', function() {
@@ -114,8 +127,9 @@ gulp.task('browserSync', function() {
 })
 
 // Dev task with browserSync
-gulp.task('dev', ['browserSync', 'sass', 'minify-css', 'minify-js'], function() {
-  gulp.watch('scss/*.scss', ['sass']);
+gulp.task('dev', ['browserSync', 'sass', 'minify-css', 'minify-js', 'pug'], function() {
+  gulp.watch('scss/**/*', ['sass']);
+  gulp.watch('pug/**/*', ['pug']);
   gulp.watch('css/*.css', ['minify-css']);
   gulp.watch('js/*.js', ['minify-js']);
   // Reloads the browser whenever HTML or JS files change
