+++ pug/charts.pug
@@ -1,0 +1,73 @@
+doctype html
+html(lang='en')
+
+  head
+    meta(charset='utf-8')
+    meta(http-equiv='X-UA-Compatible', content='IE=edge')
+    meta(name='viewport', content='width=device-width, initial-scale=1, shrink-to-fit=no')
+    meta(name='description', content='')
+    meta(name='author', content='')
+
+    title SB Admin - Start Bootstrap Template
+
+    include includes/css/core.pug
+
+    include includes/css/custom.pug
+
+  body#page-top.fixed-nav.sticky-footer.bg-dark
+    include includes/navbar.pug
+    .content-wrapper
+      .container-fluid
+        // Breadcrumbs
+        ol.breadcrumb
+          li.breadcrumb-item
+            a(href='#') Dashboard
+          li.breadcrumb-item.active Charts
+        // Area Chart Example
+        .card.mb-3
+          .card-header
+            i.fa.fa-area-chart
+            |  Area Chart Example
+          .card-body
+            canvas#myAreaChart(width='100%', height='30')
+          .card-footer.small.text-muted
+            | Updated yesterday at 11:59 PM
+        .row
+          .col-lg-8
+            // Example Bar Chart Card
+            .card.mb-3
+              .card-header
+                i.fa.fa-bar-chart
+                |  Bar Chart Example
+              .card-body
+                canvas#myBarChart(width='100', height='50')
+              .card-footer.small.text-muted
+                | Updated yesterday at 11:59 PM
+          .col-lg-4
+            // Example Pie Chart Card
+            .card.mb-3
+              .card-header
+                i.fa.fa-pie-chart
+                |  Pie Chart Example
+              .card-body
+                canvas#myPieChart(width='100%', height='100')
+              .card-footer.small.text-muted
+                | Updated yesterday at 11:59 PM
+      // /.container-fluid
+      // /.content-wrapper
+
+      include includes/footer.pug
+
+      include includes/scroll-to-top.pug
+
+      include includes/modals/logout.pug
+
+      include includes/js/core.pug
+
+      // Page level plugin JavaScript
+      script(src='vendor/chart.js/Chart.min.js')
+
+      include includes/js/custom.pug
+
+      // Custom scripts for this page
+      script(src='js/sb-admin-charts.min.js')
