+++ pug/forgot-password.pug
@@ -1,0 +1,35 @@
+doctype html
+html(lang='en')
+
+  head
+    meta(charset='utf-8')
+    meta(http-equiv='X-UA-Compatible', content='IE=edge')
+    meta(name='viewport', content='width=device-width, initial-scale=1, shrink-to-fit=no')
+    meta(name='description', content='')
+    meta(name='author', content='')
+
+    title SB Admin - Start Bootstrap Template
+
+    include includes/css/core.pug
+
+    include includes/css/custom.pug
+
+  body.bg-dark
+    .container
+      .card.card-login.mx-auto.mt-5
+        .card-header
+          | Reset Password
+        .card-body
+          .text-center.mt-4.mb-5
+            h4 Forgot your password?
+            p
+              | Enter your email address and we will send you instructions on how to reset your password.
+          form
+            .form-group
+              input#exampleInputEmail1.form-control(type='email', aria-describedby='emailHelp', placeholder='Enter email address')
+            a.btn.btn-primary.btn-block(href='login.html') Reset Password
+          .text-center
+            a.d-block.small.mt-3(href='register.html') Register an Account
+            a.d-block.small(href='login.html') Login Page
+
+    include includes/js/core.pug
