+++ pug/includes/css/core.pug
@@ -1,0 +1,4 @@
+// Bootstrap core CSS
+link(href='vendor/bootstrap/css/bootstrap.min.css', rel='stylesheet')
+// Custom fonts for this template
+link(href='vendor/font-awesome/css/font-awesome.min.css', rel='stylesheet', type='text/css')
