+++ pug/includes/css/custom.pug
@@ -1,0 +1,2 @@
+// Custom styles for this template
+link(href='css/sb-admin.css', rel='stylesheet')
