+++ pug/includes/footer.pug
@@ -1,0 +1,4 @@
+footer.sticky-footer
+  .container
+    .text-center
+      small Copyright © Your Website 2017
