+++ pug/includes/js/core.pug
@@ -1,0 +1,6 @@
+// Bootstrap core JavaScript
+script(src='vendor/jquery/jquery.min.js')
+script(src='vendor/popper/popper.min.js')
+script(src='vendor/bootstrap/js/bootstrap.min.js')
+// Core plugin JavaScript
+script(src='vendor/jquery-easing/jquery.easing.min.js')
