+++ pug/includes/js/custom.pug
@@ -1,0 +1,2 @@
+// Custom scripts for all pages
+script(src='js/sb-admin.min.js')
