+++ pug/includes/scroll-to-top.pug
@@ -1,0 +1,3 @@
+// Scroll to Top Button
+a.scroll-to-top.rounded(href='#page-top')
+  i.fa.fa-angle-up
