+++ scss/_global.scss
@@ -5,6 +5,10 @@ html {
   min-height: 100%;
 }
 
+body {
+  overflow-x: hidden;
+}
+
 body.sticky-footer {
   margin-bottom: 56px;
   .content-wrapper {
