+++ scss/navbar/_navbar_static.scss
