+++ scss/sb-admin.scss
@@ -2,10 +2,11 @@
 @import "mixins.scss";
 @import "global.scss";
 @import "utilities.scss";
+@import "navbar/navbar_global.scss";
+@import "navbar/navbar_fixed.scss";
+@import "navbar/navbar_static.scss";
+@import "navbar/navbar_toggle.scss";
+@import "navbar/navbar_colors.scss";
 @import "cards.scss";
-@import "navbar.scss";
-@import "navbar_fixed.scss";
-@import "navbar_static.scss";
-@import "navbar_toggle.scss";
 @import "login.scss";
 @import "footer.scss";
